+++ Application/src/Solver/Days/SolutionDay01.cs
@@ -4,10 +4,10 @@ namespace Application.Solver.Days;
 
 public class SolutionDay01() : DaySolver(1)
 {
-    public override int Resolve(int part, string input)
+    public override ulong Resolve(int part, string input)
     {
         var (left, right) = SplitInLeftRightLists(input);
-        return part == 1 ? Part1(left, right) : Part2(left, right);
+        return (ulong)(part == 1 ? Part1(left, right) : Part2(left, right));
     }
 
     protected static (List<int>, List<int>) SplitInLeftRightLists(string input)
